Average distance and flyers for each teams
In Graph, added 2 new arguments: the average distance to travel for each
teams and the average amout of flyers to deliver.
Then I modified the constructor with these new parameter and with a
default value of 0. I hope it will for Armand's subgraph division.
(I don't compute the average in constructor beceause we don't want to
compute that for each subgraph)